refactor: injector service
